Move out memoryMap implementation details and definitions from header file
